#331 merge private and group
