+++ back/routes/calendar.js
@@ -1,6 +1,6 @@
 const express = require("express");
 
-const { sequelize, ProfileImage } = require("../models");
+const { sequelize, ProfileImage, ChildEvent } = require("../models");
 const { User } = require("../models");
 const { Calendar } = require("../models");
 const { CalendarMember } = require("../models");
@@ -15,12 +15,7 @@ router.get("/getMyCalendars", authJWT, async (req, res, next) => {
   try {
     const me = await User.findOne({ where: { id: req.myId } });
 
-    const privateCalendar = await me.getPrivateCalendar();
-
-    const groupCalendars = await me.getGroupCalendars({
-      attributes: {
-        exclude: ["OwnerId"],
-      },
+    const myCalendars = await me.getCalendars({
       include: [
         {
           model: User,
@@ -43,10 +38,7 @@ router.get("/getMyCalendars", authJWT, async (req, res, next) => {
       joinTableAttributes: ["authority"],
     });
 
-    return res.status(200).send({
-      privateCalendar: privateCalendar,
-      groupCalendars: groupCalendars,
-    });
+    return res.status(200).send(myCalendars);
   } catch (error) {
     console.error(error);
     next(error);
@@ -86,7 +78,7 @@ router.post("/createGroupCalendar", authJWT, async (req, res, next) => {
   }
 });
 
-router.post("/editGroupCalendar", authJWT, async (req, res, next) => {
+router.post("/editCalendar", authJWT, async (req, res, next) => {
   try {
     const myAuthority = await CalendarMember.findOne({
       where: {
@@ -114,30 +106,31 @@ router.post("/editGroupCalendar", authJWT, async (req, res, next) => {
           transaction: t,
         }
       );
-
-      const members = await changeCalendar.getCalendarMembers();
-      await Promise.all(
-        members.map((member) =>
-          Alert.create(
-            {
-              UserId: member.id,
-              type: "calenderChanged",
-              content: `${changeCalendar.name}캘린더가 수정되었습니다. 확인해주세요!`,
-            },
-            { transaction: t }
-          )
-        )
-      );
-      return res.status(200).send(changeCalendar);
     });
+
+    return res.status(200).send(changeCalendar);
   } catch (error) {
     console.error(error);
     next(error);
   }
 });
 
-router.post("/deleteGroupCalendar", authJWT, async (req, res, next) => {
+router.post("/deleteCalendar", authJWT, async (req, res, next) => {
   try {
+    const calendar = await Calendar.findOne({
+      where: {
+        [Op.and]: {
+          id: req.body.calendarId,
+        },
+      },
+    });
+
+    if (calendar.private) {
+      return res
+        .status(400)
+        .send({ message: "개인 캘린더는 삭제할 수 없습니다!" });
+    }
+
     const myAuthority = await CalendarMember.findOne({
       where: {
         [Op.and]: {
@@ -148,28 +141,14 @@ router.post("/deleteGroupCalendar", authJWT, async (req, res, next) => {
     });
 
     if (myAuthority.authority < 3) {
-      return res.status(400).send({ message: "권한이 없습니다!" });
+      return res.status(402).send({ message: "권한이 없습니다!" });
     }
 
-    const exCalendar = await Calendar.findOne({
-      where: {
-        id: req.body.calendarId,
-      },
-    });
-    if (!exCalendar) {
-      return res.status(402).send({
-        message:
-          "삭제하려는 캘린더를 찾을 수 없습니다 입력값을 다시 확인해주세요",
+    await sequelize.transaction(async (t) => {
+      await calendar.destroy({
+        force: true,
+        transaction: t,
       });
-    }
-
-    await deleteAlertsByCalendarId(req.body.calendarId);
-
-    await Calendar.destroy({
-      where: {
-        id: req.body.calendarId,
-      },
-      force: true,
     });
 
     return res.status(200).send({ success: true });
@@ -189,7 +168,7 @@ router.post("/inviteGroupCalendar", authJWT, async (req, res, next) => {
         },
       },
     });
-    if (myAuthority.authority < 3) {
+    if (myAuthority?.authority < 3) {
       return res.status(400).send({ message: "권한이 없습니다!" });
     }
 
@@ -209,6 +188,7 @@ router.post("/inviteGroupCalendar", authJWT, async (req, res, next) => {
         },
       },
     });
+
     if (alreadyInvite) {
       return res
         .status(403)
@@ -239,6 +219,12 @@ router.post("/inviteGroupCalendar", authJWT, async (req, res, next) => {
         where: { id: req.body.calendarId },
       });
 
+      if (InviteCalendar?.private) {
+        return res
+          .status(406)
+          .send({ message: "개인 캘린더에는 초대불가능 합니다!" });
+      }
+
       await Alert.create(
         {
           UserId: guest.id,
@@ -267,18 +253,41 @@ router.post("/resignCalendar", authJWT, async (req, res, next) => {
       return res.status(400).send({ message: "존재하지 않는 캘린더 입니다!" });
     }
 
+    if (groupCalendar.private) {
+      return res
+        .status(402)
+        .send({ message: "개인 캘린더는 탈퇴 할 수 없습니다!" });
+    }
+
     await sequelize.transaction(async (t) => {
-      await CalendarMember.destroy(
-        {
-          where: {
-            [Op.and]: { UserId: req.myId, CalendarId: req.body.calendarId },
+      await CalendarMember.destroy({
+        where: {
+          [Op.and]: { UserId: req.myId, CalendarId: req.body.calendarId },
+        },
+        transaction: t,
+        force: true,
+      });
+
+      const myCalendar = await Calendar.findOne({
+        where: {
+          [Op.and]: {
+            private: true,
+            OwnerId: req.myId,
           },
         },
-        { transaction: t },
-        {
+      });
+
+      // childEvent에서 해당 캘린더에 대한 이벤트 모두 지워줘야함
+      await ChildEvent.destroy({
+        where: {
+          [Op.and]: {
+            privateCalendarId: myCalendar.id,
+            originCalendarId: req.body.calendarId,
+          },
+          transaction: t,
           force: true,
-        }
-      );
+        },
+      });
     });
 
     return res.status(200).send({ success: true });
@@ -323,15 +332,24 @@ router.post("/sendOutUser", authJWT, async (req, res, next) => {
     }
 
     await sequelize.transaction(async (t) => {
-      await CalendarMember.destroy(
-        {
-          where: {
-            [Op.and]: { UserId: member.id, CalendarId: req.body.calendarId },
-          },
+      await CalendarMember.destroy({
+        where: {
+          [Op.and]: { UserId: req.myId, CalendarId: req.body.calendarId },
         },
-        { transaction: t },
-        { force: true }
-      );
+        transaction: t,
+        force: true,
+      });
+
+      await ChildEvent.destroy({
+        where: {
+          [Op.and]: {
+            privateCalendarId: myCalendar.id,
+            originCalendarId: req.body.calendarId,
+          },
+          transaction: t,
+          force: true,
+        },
+      });
     });
 
     return res.status(200).send({ success: true });
@@ -366,7 +384,7 @@ router.post("/acceptCalendarInvite", authJWT, async (req, res, next) => {
       return res.status(402).send({ message: "이미 캘린더의 그룹원 입니다!" });
     }
 
-    const deletedInvite = await Invite.findOne({
+    const invitation = await Invite.findOne({
       where: {
         [Op.and]: {
           guestId: req.myId,
@@ -375,14 +393,14 @@ router.post("/acceptCalendarInvite", authJWT, async (req, res, next) => {
       },
     });
 
-    if (!deletedInvite) {
+    if (!invitation) {
       return res.status(403).send({ message: "존재하지 않는 초대장입니다!" });
     }
 
     await sequelize.transaction(async (t) => {
       await CalendarMember.create(
         {
-          authority: deletedInvite.authority,
+          authority: invitation.authority,
           UserId: req.myId,
           CalendarId: req.body.calendarId,
         },
@@ -455,7 +473,7 @@ router.post("/rejectCalendarInvite", authJWT, async (req, res, next) => {
       return res.status(402).send({ message: "이미 캘린더의 그룹원 입니다!" });
     }
 
-    const deletedInvite = await Invite.findOne({
+    const invitation = await Invite.findOne({
       where: {
         [Op.and]: {
           guestId: req.myId,
@@ -463,14 +481,14 @@ router.post("/rejectCalendarInvite", authJWT, async (req, res, next) => {
         },
       },
     });
-    if (!deletedInvite) {
+    if (!invitation) {
       return res.status(403).send({ message: "존재하지 않는 초대장입니다!" });
     }
 
     await sequelize.transaction(async (t) => {
       await Alert.create(
         {
-          UserId: deletedInvite.CalendarHostId,
+          UserId: invitation.hostId,
           type: "calendarInviteReject",
           content: `${me.nickname}님이 ${groupCalendar.name} 캘린더의 초대를 거부하셨습니다.`,
         },
@@ -511,7 +529,6 @@ router.post("/giveAuthority", authJWT, async (req, res, next) => {
       },
     });
 
-    console.log(myAuthority.authority);
     if (myAuthority.authority < 3) {
       return res.status(400).send({ message: "권한이 없습니다!" });
     }
@@ -527,8 +544,14 @@ router.post("/giveAuthority", authJWT, async (req, res, next) => {
       where: { id: req.body.calendarId },
     });
 
+    if (groupCalendar.private) {
+      return res
+        .status(403)
+        .send({ message: "개인 캘린더의 권한은 수정할 수 없습니다!" });
+    }
+
     if (!groupCalendar) {
-      return res.status(403).send({ message: "존재하지 않는 캘린더입니다!" });
+      return res.status(405).send({ message: "존재하지 않는 캘린더입니다!" });
     }
 
     const isGroupMember = await CalendarMember.findOne({
