+++ back/swagger/calendarAPI.yaml
@@ -12,18 +12,7 @@
         description: 성공
         schema:
           properties:
-            privateCalendar:
-              type: object
-              properties:
-                id:
-                  type: integer
-                name:
-                  type: string
-                color:
-                  type: string
-                UserId:
-                  type: string
-            groupCalendars:
+            calendars:
               type: array
               items:
                 type: object
@@ -100,11 +89,11 @@
         $ref: "./openapi.yaml#/returnCalendar"
       500:
         $ref: "./openapi.yaml#/serverError"
-/api/calendar/editGroupCalendar:
+/api/calendar/editCalendar:
   post:
     tags:
       - calendar
-    summary: 그룹캘린더 수정
+    summary: 그룹캘린더나 개인캘린더 수정
     parameters:
       - name: Authorization
         in: header
@@ -155,20 +144,21 @@
       200:
         $ref: "./openapi.yaml#/success"
       400:
-        description: 권한을 부여하는 유저가 권한이 3 미만인 경우
+        description: 개인 캘린더를 삭제하려는 경우
+        schema:
+          type: object
+          properties:
+            message:
+              type: string
+              example: 개인 캘린더는 삭제할 수 없습니다!
+      402:
+        description: 캘린더를 삭제하려는 유저가 권한이 3 미만인 경우
         schema:
           type: object
           properties:
             message:
               type: string
               example: 권한이 없습니다!
-      402:
-        description: 삭제하려는 캘린더가 존재하지 않는 경우
-        schema:
-          properties:
-            message:
-              type: string
-              example: 삭제하려는 캘린더를 찾을 수 없습니다 입력값을 다시 확인해주세요
       500:
         $ref: "./openapi.yaml#/serverError"
 /api/calendar/inviteGroupCalendar:
@@ -227,6 +217,14 @@
             message:
               type: string
               example: 이미 캘린더의 그룹원 입니다!
+      406:
+        description: 개인 캘린더에 초대하려고 할 경우
+        schema:
+          type: object
+          properties:
+            message:
+              type: string
+              example: 개인 캘린더에는 초대불가능 합니다!
       500:
         $ref: "./openapi.yaml#/serverError"
 /api/calendar/acceptCalendarInvite:
@@ -329,7 +327,7 @@
   post:
     tags:
       - calendar
-    summary: 캘린더 탈퇴하기
+    summary: 그룹 캘린더 탈퇴하기
     parameters:
       - name: Authorization
         in: header
@@ -353,6 +351,14 @@
             message:
               type: string
               example: 존재하지 않는 캘린더 입니다!
+      402:
+        description: 개인 캘린더를 탈퇴하려 할 때
+        schema:
+          type: object
+          properties:
+            message:
+              type: string
+              example: 개인 캘린더는 탈퇴 할 수 없습니다!
       500:
         $ref: "./openapi.yaml#/serverError"
 /api/calendar/sendOutUser:
