--- back/swagger/privateCalendarAPI.yaml
@@ -1,103 +1,0 @@
-/api/privateCalendar/editPrivateCalendar:
-  post:
-    tags:
-      - privateCalendar
-    summary: 개인 캘린더 수정
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            calendarId:
-              type: integer
-            calendarName:
-              type: string
-            calendarColor:
-              type: string
-    responses:
-      200:
-        description: 성공
-        schema:
-          properties:
-            id:
-              type: integer
-            name:
-              type: string
-            color:
-              type: string
-            UserId:
-              type: integer
-      400:
-        description: 개인 캘린더 조회 결과가 없는 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 개인 캘린더 조회 결과가 없습니다 입력값을 다시 확인해주세요
-      402:
-        description: 수정 시도한 유저가 개인 캘린더 주인이 아닐 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 수정 시도하는 유저가 개인 캘린더의 주인이 아닙니다
-      500:
-        $ref: "./openapi.yaml#/serverError"
-/api/privateCalendar/editPrivateCalendarColor:
-  post:
-    tags:
-      - privateCalendar
-    summary: 개인 캘린더 수정
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            calendarId:
-              type: integer
-            calendarColor:
-              type: string
-    responses:
-      200:
-        description: 성공
-        schema:
-          properties:
-            id:
-              type: integer
-            name:
-              type: string
-            color:
-              type: string
-            UserId:
-              type: integer
-      400:
-        description: 개인 캘린더 조회 결과가 없는 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 개인 캘린더 조회 결과가 없습니다 입력값을 다시 확인해주세요
-      402:
-        description: 수정 시도한 유저가 개인 캘린더 주인이 아닐 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 수정 시도하는 유저가 개인 캘린더의 주인이 아닙니다
-      500:
-        $ref: "./openapi.yaml#/serverError"
-$ref: "./privateEventAPI.yaml"
