--- back/swagger/privateEventAPI.yaml
@@ -1,257 +1,0 @@
-/api/privateEvent/getGroupEvent:
-  post:
-    tags:
-      - privateEvent
-    summary: 특정 개인 이벤트의 모든 정보 가져오기
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            eventId:
-              type: integer
-    responses:
-      200:
-        description: 성공
-        schema:
-          properties:
-            events:
-              type: object
-              properties:
-                id:
-                  type: integer
-                color:
-                  type: string
-                priority:
-                  type: string
-                memo:
-                  type: string
-                startTime:
-                  type: string
-                endTime:
-                  type: string
-                groupEventId:
-                  type: integer
-                PrivateCalendarId:
-                  type: integer
-                EventHost:
-                  type: object
-                  properties:
-                    id:
-                      type: integer
-                    email:
-                      type: string
-                    nickname:
-                      type: string
-                EventMembers:
-                  type: array
-                  items:
-                    type: object
-                    properties:
-                      id:
-                        type: string
-                      email:
-                        type: string
-                      nickname:
-                        type: string
-                      ProfileImages:
-                        type: object
-                        properties:
-                          src:
-                            type: integer
-                      EventMember:
-                        type: object
-                        properties:
-                          state:
-                            type: integer
-                          UserId:
-                            type: integer
-                          EventId:
-                            type: integer
-            realTimeAlert:
-              type: object
-              properties:
-                type:
-                  type: string
-                  example: minute, hour, day, week / day, week (allDay)
-                time:
-                  type: integer
-                  example: 1
-                hour:
-                  type: integer
-                  example: 1 (allDay)
-                minute:
-                  type: integer
-                  example: 1 (allDay)
-      500:
-        $ref: "./openapi.yaml#/serverError"
-/api/privateEvent/createPrivateEvent:
-  post:
-    tags:
-      - privateEvent
-    summary: 개인 이벤트 만들기
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            eventId:
-              type: integer
-              example: 1
-            eventName:
-              type: string
-            color:
-              type: integer
-              example: "#0e0e0e"
-            busy:
-              type: integer
-            memo:
-              type: string
-            startTime:
-              type: string
-              example: "2022-06-02T00:00:00"
-            endTime:
-              type: string
-              example: "2022-06-02T00:00:00"
-            allDay:
-              type: integer
-              example: 1
-            alerts:
-              type: array
-              items:
-                type: object
-                properties:
-                  type:
-                    type: string
-                    example: "minute / hour / day / week"
-                  time:
-                    type: integer
-    responses:
-      200:
-        $ref: "./openapi.yaml#/returnPrivateEvent"
-      500:
-        $ref: "./openapi.yaml#/serverError"
-/api/privateEvent/editPrivateEvent:
-  post:
-    tags:
-      - privateEvent
-    summary: 개인 이벤트 수정
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            eventId:
-              type: integer
-              example: 1
-            eventName:
-              type: string
-            color:
-              type: integer
-              example: "#0e0e0e"
-            busy:
-              type: integer
-            memo:
-              type: string
-            startTime:
-              type: string
-              example: "2022-06-02T00:00:00"
-            endTime:
-              type: string
-              example: "2022-06-02T00:00:00"
-            allDay:
-              type: integer
-              example: 1
-            alerts:
-              type: array
-              items:
-                type: object
-                properties:
-                  type:
-                    type: string
-                    example: "minute / hour / day / week"
-                  time:
-                    type: integer
-    responses:
-      200:
-        $ref: "./openapi.yaml#/returnPrivateEvent"
-      400:
-        description: 수정할 개인 이벤트 조회 결과가 없을 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 수정할 개인이벤트의 조회 결과가 없습니다
-      500:
-        $ref: "./openapi.yaml#/serverError"
-/api/privateEvent/editPrivateEventColor:
-  post:
-    tags:
-      - privateEvent
-    summary: 개인 이벤트 색상 수정
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            eventId:
-              type: integer
-            color:
-              type: string
-    responses:
-      200:
-        $ref: "./openapi.yaml#/returnPrivateEvent"
-      400:
-        description: 수정할 개인 이벤트 조회 결과가 없을 경우
-        schema:
-          type: object
-          properties:
-            message:
-              type: string
-              example: 수정할 개인이벤트의 조회 결과가 없습니다
-      500:
-        $ref: "./openapi.yaml#/serverError"
-/api/privateEvent/deletePrivateEvent:
-  post:
-    tags:
-      - privateEvent
-    summary: 개인 이벤트 삭제
-    parameters:
-      - name: Authorization
-        in: header
-        required: true
-      - name: body
-        in: body
-        required: true
-        schema:
-          type: object
-          properties:
-            eventId:
-              type: integer
-    responses:
-      200:
-        $ref: "./openapi.yaml#/success"
-      500:
-        $ref: "./openapi.yaml#/serverError"
-$ref: "./calendarAPI.yaml"
